로켓 이펙트 업데이트 및 세레모니 전 이펙트 업데이트
