+++ components/Game/mypuzzle.tsx
@@ -20,9 +20,8 @@ import MakeVideoTwirl from "../FaceDetection/MakeVideoTwirl";
 import MyTwirlParticles from "../PageElements/Particles/mytwirlParticles";
 import MyRocketParticles from "../PageElements/Particles/myrocketParticles";
 import MakeVideoLip from "../FaceDetection/MakeVideoLip";
+import CeremonyParticles from "../PageElements/Particles/ceremonyParticles";
 
-let isRightPlace: boolean[] = [false, false, false, false, false, false, false, false, false];
-let i: number;
 
 // import Segment from './Segment'
 const PuzzleSegment = dynamic(import("@/components/Game/Segment"), {
@@ -36,6 +35,7 @@ interface Props {
     dataChannel: RTCDataChannel | undefined;
 }
 const fanFareSoundUrl = '/sounds/Fanfare.mp3';
+const winSoundUrl = '/sounds/YouWin.mp3';
 
 
 function MyPuzzle({ auth, videoId, dataChannel }: Props) {
@@ -50,6 +50,8 @@ function MyPuzzle({ auth, videoId, dataChannel }: Props) {
     const puzzleCompleteCounter = useSelector((state: any) => state.puzzleComplete);
     const router = useRouter();
     const [fanFareSoundPlay] = useSound(fanFareSoundUrl);
+    const [winSoundPlay] = useSound(winSoundUrl);
+    const [isFinished, setIsFinished] = useState(false);
 
     //dataChannel에 addEventListner 붙이기 (하나의 dataChannel에 이벤트리스너를 여러번 붙이는 것은 문제가 없다.)
 
@@ -102,19 +104,24 @@ function MyPuzzle({ auth, videoId, dataChannel }: Props) {
 
     useEffect(() => {
         if (puzzleCompleteCounter.mine === 9) {
-            const myface = document.getElementById("myface");
-            myface!.style.display = "block";
-            document.getElementById("fullscreen")!.style.display = "none";
-            document.getElementById("itembar")!.style.display = "none";
-            document.getElementById("face")!.style.display = "block";
-            fanFareSoundPlay();
+            setIsFinished(true)
+            winSoundPlay();
             setTimeout(() => {
-                router
-                    .replace({
-                        pathname: "/ready",
-                    })
-                    .then(() => router.reload());
-            }, 15000);
+                const myface = document.getElementById("myface");
+                myface!.style.display = "block";
+                document.getElementById("fullscreen")!.style.display = "none";
+                document.getElementById("itembar")!.style.display = "none";
+                document.getElementById("face")!.style.display = "block";
+                fanFareSoundPlay();
+                setTimeout(() => {
+                    router
+                        .replace({
+                            pathname: "/ready",
+                        })
+                        .then(() => router.reload());
+                }, 15000);
+            }, 5000)
+
         }
     }, [puzzleCompleteCounter.mine]);
 
@@ -127,6 +134,10 @@ function MyPuzzle({ auth, videoId, dataChannel }: Props) {
 
     return (
         <>
+            {isFinished && <>
+                <div className="fixed ml-[50vw] mt-[270px] w-[100vw] text-center overflow-visible text-9xl z-50 text-red-900"> YOU WIN </div>
+                <CeremonyParticles />
+            </>}
             {
                 [...Array(9)].map((_, i) => {
                     return (
@@ -139,20 +150,26 @@ function MyPuzzle({ auth, videoId, dataChannel }: Props) {
                 }
                 )
             }
+            {isFinished &&
+                <>
+                    <div className={`absolute flex justify-center items-center w-[640px] h-[480px] mt-[100px] ${styles.finish}`}>
+                        <img src="../images/finish.gif" className={`z-40 ${styles.gif}`} draggable="false" style={{ pointerEvents: "none" }} />
+                    </div>
+                </>
+            }
+
             {/* 아이템 쓸 때 나오는 효과 */}
             <div className="absolute grid w-[640px] h-[480px] mt-[100px]" style={{ pointerEvents: "none" }}>
                 {mySegmentState.segementState === 'ice' && (<div id="myice" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyIceFlakeParticles /> <img src="../images/icemine.gif" className={`z-50 ${styles.gif}`} draggable="false" style={{ pointerEvents: "none" }} /> </div>)}
                 {mySegmentState.segementState === 'magnet' && (<div id="mymagnet" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyBlackhallParticles /> <img src="../images/blackholemine.gif" className={`z-50 ${styles.gif}`} draggable="false" style={{ pointerEvents: "none" }} /> </div>)}
                 {mySegmentState.segementState === 'lip' && (<div id="mylip" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyLipParticles /> <img src="../images/lipmine.gif" className={`z-50 ${styles.gif2}`} draggable="false" style={{ pointerEvents: "none" }} /> </div>)}
                 {mySegmentState.segementState === 'twirl' && (<div id="mytwirl" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyTwirlParticles />  </div>)}
-                {mySegmentState.segementState === 'rocket' && (<div id="myrocket" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyRocketParticles />  </div>)}
+                {mySegmentState.segementState === 'rocket' && (<div id="myrocket" className={`flex fill`} style={{ pointerEvents: "none" }} > <MyRocketParticles /> <img src="../images/rocketmine.gif" className={`z-50 ${styles.gif}`} draggable="false" style={{ pointerEvents: "none" }} /> </div> )}
             </div>
             <MyBar score={puzzleCompleteCounter.mine} />
             <Modal segmentState={mySegmentState.segementState} />
             {mySegmentState.segementState === 'lip' && <MakeVideoLip auth={auth} />}
             {mySegmentState.segementState === 'twirl' && <MakeVideoTwirl videoId={videoId} auth={auth} />}
-
-
         </>
     );
 }
